test for the the existance of test.log.eps before trying to convert the file (#18)
* added test for existance of logs/test.log.eps before html is generated

if the file test.log.eps does not exist, the following instructions will
fail. So the code assumes that test.log.eps exists. When you run train.lua
the first time, this is not true. I think test.log.eps is not generated when
not enough data is there, so for the first epoch. Lateron it will be
there.

* whitespace only change after testing for logs/test.log.eps

indent after testing for logs/test.log.eps